requirements.txt
async-timeout==4.0.2
certifi==2022.12.7
click==8.1.3
clickhouse-connect==0.4.7
Deprecated==1.2.13
elastic-transport==8.4.0
elasticsearch==7.17.8
elasticsearch-dsl==7.4.0
Flask==2.1.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
psycopg==3.0.12
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-dateutil==2.8.2
redis==4.2.2
six==1.16.0
sqlparse==0.4.2
tabulate==0.8.10
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.14.1
yoyo-migrations==7.3.2
zipp==3.8.1